Rename test.cpp -> matrix_template.hpp & create test.cpp
Contains major updates to the matrix template, including completion of the matrix implementations to a testable degree, with addition and some forms of multiplication.
There is a bug to note within the _matrix (uint8_t) specialization [that is all that is known currently], where the destroy function call on a rvalue object causes strange behavior and a possible crash.